Made it all longgggg also default goal

+++ src/main/java/io/github/seggan/sfcalc/CalcHandler.java
@@ -68,7 +68,7 @@ public class CalcHandler implements Listener {
                 return;
             }
 
-            Calculator.printResults(Calculator.calculate(item, plugin), e.getWhoClicked(), "sfcalc", item, 1, plugin);
+            Calculator.printResults(Calculator.calculate(item, plugin), e.getWhoClicked(), "sfcalc", item, (long) 1, plugin);
         }
     }
 }
